split seq_region rank checking away from main SeqRegionsTopLevel check as this can fail for EG dbs where we use 1 for chromosome, which not all genomes have
